+++ pool/app/pool_common/locales/locales_de.ml
@@ -250,6 +250,12 @@ let warning_to_string : warning -> string = function
 ;;
 
 let rec error_to_string = function
+  | AccountTemporarilySuspended ptime ->
+    ptime
+    |> Utils.Ptime.formatted_date_time
+    |> Format.asprintf
+         "Zu viele fehlgeschlagene Anmeldeversuche. Diese E-Mail-Adresse ist \
+          gesperrt, bis %s"
   | AccessDenied -> "Zugriff verweigert"
   | AccessDeniedMessage ->
     "Der Zugriff auf die gewünschte Seite ist nicht möglich."
@@ -377,7 +383,15 @@ let rec error_to_string = function
       (err2 |> error_to_string |> CCString.uncapitalize_ascii)
   | PasswordConfirmationDoesNotMatch ->
     "Passwortbestätigung stimmt nicht mit dem neuen Passwort überein."
-  | PasswordPolicy -> "Passwort stimmt nicht mit der benötigten Policy überein!"
+  | PasswordPolicyMinLength n ->
+    Format.asprintf "Das Passwort muss mindestens %i Zeichen lang sein." n
+  | PasswordPolicyCapitalLetter ->
+    "Das Passwort muss einen Grossbuchstaben enthalten."
+  | PasswordPolicyNumber -> "Das Passwort muss eine Zahl enthalten."
+  | PasswordPolicySpecialChar chars ->
+    Format.asprintf
+      "Das Passwort muss eines der folgenden Zeichen enthalten: %s"
+      (chars |> CCList.map CCString.of_char |> CCString.concat " ")
   | PasswordResetFailMessage ->
     "Falls ein Account zu der von dir eingegebenen E-Mail Adresse existiert,  \
      wird dir ein E-Mail mit einem Link zur Passwort zurücksetzung gesendet."