+++ pool/app/pool_user/entity.ml
@@ -17,9 +17,92 @@ module PasswordConfirmed = struct
 end
 
 module Password = struct
+  module Policy = struct
+    type rule =
+      | MinLength of int
+      | MustContainCapitalLetter
+      | MustContainNumber
+      | MustContainSpecialChar of char list
+
+    type t = rule list
+
+    let validate_min_length num p =
+      if CCString.length p < num
+      then Error (PoolError.PasswordPolicyMinLength num)
+      else Ok p
+    ;;
+
+    let validate_characters validator error p =
+      p
+      |> CCString.to_list
+      |> CCList.fold_left (fun is_ok c -> is_ok || validator c) false
+      |> function
+      | true -> Ok p
+      | false -> Error error
+    ;;
+
+    let validate_capital_letter =
+      let validate c = CCChar.to_int c >= 65 && CCChar.to_int c <= 90 in
+      validate_characters validate PoolError.PasswordPolicyCapitalLetter
+    ;;
+
+    let validate_number =
+      let validate c = CCChar.to_int c >= 48 && CCChar.to_int c <= 57 in
+      validate_characters validate PoolError.PasswordPolicyNumber
+    ;;
+
+    let validate_special_char chars p =
+      chars
+      |> CCList.fold_left (fun is_ok c -> is_ok || CCString.contains p c) false
+      |> function
+      | true -> Ok p
+      | false -> Error (PoolError.PasswordPolicySpecialChar chars)
+    ;;
+
+    let default_special_char_set =
+      [ '!'
+      ; '?'
+      ; '*'
+      ; '+'
+      ; '-'
+      ; '_'
+      ; '&'
+      ; '%'
+      ; '('
+      ; ')'
+      ; '}'
+      ; '{'
+      ; '$'
+      ; ','
+      ; '.'
+      ]
+    ;;
+
+    let default_policy =
+      [ MinLength 8
+      ; MustContainCapitalLetter
+      ; MustContainNumber
+      ; MustContainSpecialChar default_special_char_set
+      ]
+    ;;
+
+    let valdate password =
+      let open CCResult in
+      CCList.fold_left
+        (fun password rule ->
+          password
+          >>=
+          match rule with
+          | MinLength n -> validate_min_length n
+          | MustContainCapitalLetter -> validate_capital_letter
+          | MustContainNumber -> validate_number
+          | MustContainSpecialChar chars -> validate_special_char chars)
+        (Ok password)
+    ;;
+  end
+
   type t = string [@@deriving eq]
 
-  let create password = Ok password
   let to_sihl m = m
   let show m = CCString.repeat "*" @@ CCString.length m
 
@@ -27,17 +110,15 @@ module Password = struct
     Format.fprintf formatter "%s" m
   ;;
 
-  let schema ?(field = PoolError.Field.Password) () =
-    Pool_common.Utils.schema_decoder create show field
-  ;;
-
-  let default_password_policy p =
-    if CCString.length p < 8 then Error PoolError.PasswordPolicy else Ok ()
-  ;;
-
-  let validate ?(password_policy = default_password_policy) password =
+  let create password =
     (* TODO: Consider checking against old password *)
-    password |> password_policy
+    Policy.valdate password Policy.default_policy
+  ;;
+
+  let create_unvalidated p = Ok p
+
+  let schema ?(field = PoolError.Field.Password) create () =
+    Pool_common.Utils.schema_decoder create show field
   ;;
 
   let validate_current_password