+++ pool/test/contact_test.ml
@@ -13,7 +13,7 @@ let check_result expected generated =
 ;;
 
 let contact_info email_address =
-  email_address, "password", "Jane", "Doe", Some Language.En
+  email_address, "Password1!", "Jane", "Doe", Some Language.En
 ;;
 
 let allowed_email_suffixes =
@@ -245,7 +245,7 @@ let update_password () =
     "john@gmail.com" |> contact_info
   in
   let contact = contact_info |> create_contact true in
-  let new_password = "testing" in
+  let new_password = "NewPassword2!" in
   let confirmation_mail = confirmation_mail contact in
   let events =
     Contact_command.UpdatePassword.(
@@ -255,10 +255,7 @@ let update_password () =
       ]
       |> decode
       |> Pool_common.Utils.get_or_failwith
-      |> handle
-           ~password_policy:(fun (_ : Pool_user.Password.t) -> Ok ())
-           contact
-           confirmation_mail)
+      |> handle contact confirmation_mail)
   in
   let expected =
     Ok
@@ -278,10 +275,48 @@ let update_password () =
   check_result expected events
 ;;
 
+let validate_password_policy password expected =
+  let res = password |> Pool_user.Password.create in
+  Alcotest.(check Test_utils.(result password error) "succeeds" expected res)
+;;
+
+let password_min_length () =
+  validate_password_policy "Pass9!" (Error (Message.PasswordPolicyMinLength 8))
+;;
+
+let password_capital_letter () =
+  validate_password_policy
+    "password9!"
+    (Error Message.PasswordPolicyCapitalLetter)
+;;
+
+let password_number () =
+  validate_password_policy "Password?" (Error Message.PasswordPolicyNumber)
+;;
+
+let password_special_char () =
+  validate_password_policy
+    "Password9"
+    (Error
+       (Message.PasswordPolicySpecialChar
+          Pool_user.Password.Policy.default_special_char_set))
+;;
+
+let valid_password () =
+  let password = "Password9*" in
+  let expected =
+    password
+    |> Pool_user.Password.create
+    |> Test_utils.get_or_failwith_pool_error
+    |> CCResult.return
+  in
+  validate_password_policy password expected
+;;
+
 let update_password_wrong_current_password () =
   let contact = "john@gmail.com" |> contact_info |> create_contact true in
-  let current_password = "something else" in
-  let new_password = "short" in
+  let current_password = "Password2!" in
+  let new_password = "Password3!" in
   let confirmation_mail = confirmation_mail contact in
   let events =
     Contact_command.UpdatePassword.(
@@ -298,11 +333,12 @@ let update_password_wrong_current_password () =
 ;;
 
 let update_password_wrong_policy () =
+  let open CCResult.Infix in
   let ((_, password, _, _, _) as contact_info) =
     "john@gmail.com" |> contact_info
   in
   let contact = contact_info |> create_contact true in
-  let new_password = "short" in
+  let new_password = "Short1!" in
   let confirmation_mail = confirmation_mail contact in
   let events =
     Contact_command.UpdatePassword.(
@@ -311,10 +347,13 @@ let update_password_wrong_policy () =
       ; Field.(PasswordConfirmation |> show), [ new_password ]
       ]
       |> decode
-      |> Pool_common.Utils.get_or_failwith
-      |> handle contact confirmation_mail)
+      >>= handle contact confirmation_mail)
   in
-  let expected = Error Message.PasswordPolicy in
+  let expected =
+    Error
+      (Message.Conformist
+         [ Field.NewPassword, Message.PasswordPolicyMinLength 8 ])
+  in
   check_result expected events
 ;;
 
@@ -323,8 +362,8 @@ let update_password_wrong_confirmation () =
     "john@gmail.com" |> contact_info
   in
   let contact = contact_info |> create_contact true in
-  let new_password = "testing" in
-  let confirmed_password = "something else" in
+  let new_password = "Password1?" in
+  let confirmed_password = "Password1*" in
   let confirmation_mail = confirmation_mail contact in
   let events =
     Contact_command.UpdatePassword.(
@@ -334,10 +373,7 @@ let update_password_wrong_confirmation () =
       ]
       |> decode
       |> Pool_common.Utils.get_or_failwith
-      |> handle
-           ~password_policy:(fun (_ : Pool_user.Password.t) -> Ok ())
-           contact
-           confirmation_mail)
+      |> handle contact confirmation_mail)
   in
   let expected = Error Pool_common.Message.PasswordConfirmationDoesNotMatch in
   check_result expected events